Containers.IsoMap

Description

Module IsoMap include type IsoMap and useful functions for working with it.

IsoMap is kind of Map with isomorphism between keys and values. So, difference between keys ans values is erased, and you can use as keys any of two available sets, and other set will be used as values.

Documentation

data IsoMap a b

size :: IsoMap a b -> Int

empty :: IsoMap a b

mapFst :: Ord b => (a -> b) -> IsoMap a c -> IsoMap b c

mapSnd :: Ord b => (a -> b) -> IsoMap c a -> IsoMap c b

mapFstRev :: Ord b => (a -> b) -> IsoMap a c -> IsoMap c b

mapSndRev :: Ord b => (a -> b) -> IsoMap c a -> IsoMap b c

module Containers.IsoMap.Iso