GRType

Description

This module represents group presentation and its sub-types.

Documentation

data A

This is a data type of GR generators.

A_Y is a generators witch obtain from alphabet letters Y of S-machine.

A_K is a generators witch forms the hubRelation.

A_R is a generators witch obtain from rules SRule of S-machine.

A_Q is a generators witch obtain from states State of S-machine.

Constructors

A_Y Y

A_K Int

A_R SRule

A_Q State

data SmbR

This is a data type of symbols in GrRelation.

Symbols can represent generators A and its invertions.

SmbA represents generators.

SmbA' represents invertions of generators.

Constructors

SmbA A

SmbA' A

data GrRelation

This is a data type of GR relations.

Relation represents relation beetwen right and left symbols SmbR.

For example, relation ab=ba will coded like

>>> Relation ([a, b], [b, a])

where a, b are SmbR

Relator is relation in witch one part is identity.

For example, ab=1 is

>>> Relator [a,b]

Constructors

Relation ([SmbR], [SmbR])

Relator [SmbR]

newtype GR

GR is a group presentation type.

It consists of set of generators A and set of relations GrRelation.

Constructors

GR (Set A, Set GrRelation)