GrammarType

Description

This module represents types of formal grammar.

In the moment we use it to represent a context-free grammar, conjunctive grammar, boolean grammar.

Documentation

newtype Terminal

Terminal is a type that represents terminal in the formal grammar Grammar.

Constructors

Terminal

Fields

terminalValue :: String

newtype Nonterminal

Nonterminal is a type that represents nonterminal in the formal grammar Grammar.

Constructors

Nonterminal

Fields

nonterminalValue :: String

data Symbol

Symbol represents symbol that can be appear in right part of the Relation.

T is for Terminal.

N is for Nonterminal.

And Eps is for empty symbol, epsilon.

Constructors

T Terminal

N Nonterminal

Eps

data Conj

Conj is helper type for representing the right part of a relation in a boolean grammar

Constructors

PosConj

Fields

symbols :: [Symbol]

NegConj

Fields

symbols :: [Symbol]

type StartSymbol = Nonterminal

This type is synonym Nonterminal and used in order to separate StartSymbol from normal Nonterminal.

data Relation

Relation is a rule of Grammar. First constructor is for working only with CFG grammars Second constructor is for working with Boolean or Conjunctive grammars, though it might be used for defining grammars too (right part of CFG relation is one PosConj in right part of boolean relation)

Constructors

Relation (Nonterminal, [Symbol])

BooleanRelation (Nonterminal, [Conj])

data GrammarType

This type we using to classify grammars.

Constructors

CFG

Conjunctive

Boolean

newtype Grammar

This type we using to represent a formal grammar.

Constructors

Grammar (Set Nonterminal, Set Terminal, Set Relation, StartSymbol)