Main

Description

This module represents interactive mode of working with program.

The user can run the project on a specified file with a grammar.

Documentation

data Options

Constructors

InputFlagString

OutputFlagString

argd :: [Arg Options]

List of options, which might be used for executing algorithm.

main :: IO ()

grammar2Group :: String -> String -> IO ()

tm2TMS :: TM -> IO ()

mainTms :: String -> String -> IO ()