SMType

Description

This module represents types of S-machine.

Documentation

data StateName

This is state name data type of State.

Every constructor here represents name of the State.

Constructors

E

X

F

P

Q

R

S

T

U

data Tag

This is tag data type of State.

Hat it's like a house roof.

Quote is a '.

Dash is a overline.

Constructors

Hat

Quote

Dash

data SMTag

SMTag is a tag for S-machines and shows what kind is it.

T4 for S-machine number 4.

T9 for number 9.

TAlpha for alpha S-machine.

TOmega for omega S-machine.

Constructors

T4

T9

TAlpha

TOmega

data TMCMD

This is data type of Turing machine commands and has uses in StateVal in order to define for which command belongs to StatevVal.

Constructors

Command [TapeCommand]

CommandAlias String

data StateVal

StateVal represents belongnes of State to tape, Turing machine commands and S-machine SM.

Constructors

StateVal

Fields

tape :: Int

tmCommand :: Maybe TMCMD

smTag :: Maybe SMTag

data State

This data type represents state of S-machine SM.

s_name represents a name of State.

s_idx represents a low index of State.

s_tags represents a set of Tag of State.

s_val represents a StateVal of State.

Constructors

State

Fields

s_name :: StateName

s_idx :: String

s_tags :: Set Tag

s_val :: Maybe StateVal

data Y

This data type represents a tape letter of S-machine SM.

Constructors

Y Square

Alpha

Delta

Omega

data Smb

This is a data type of symbols that can be on SM tape.

Symbols can be Y (SmbY), its invertions (SmbY') and State (SmbQ).

Constructors

SmbY Y

SmbY' Y

SmbQ State

newtype Word

This is a admissible word of the S-machine SM, which consist of the Smb.

Constructors

Word [Smb]

newtype SRule

This is a rule of the S-machine SM.

Left part of the pair represents what substitute.

Right part -- for what substitute.

Constructors

SRule [(Word, Word)]

data SM

This is data type of S-machine.

It consists of a lists of tape letters Y, states State and rules SRule.

Constructors

SM

Fields

yn :: [[Y]]

qn :: [Set State]

srs :: [SRule]