TMType

Description

This module represents a types of Turing machine.

Documentation

newtype State

This type represents a state of the Turing machine TM.

Constructors

State String

newtype MultiTapeStates

Type of a k-vector of tapes states, where k is a count of tapes.

Type represents all states of the Turing machine TM.

Constructors

MultiTapeStates [Set State]

newtype StartStates

Type of a k-vector of a start states of the Turing machine TM.

Constructors

StartStates [State]

newtype AccessStates

Type of a k-vector of a final states of the Turing machine TM.

Constructors

AccessStates [State]

newtype StateOmega

Type of a state in a changed Turing machine, which has a form F_q, where q is a State.

Constructors

StateOmega

Fields

state :: State

data TapeCommand

Type of a command of the Turing machine TM, which applicable on a single tape.

SingleTapeCommand is a command for normal Turing machine.

PreSMCommand is a command for Turing machine after change.

Constructors

SingleTapeCommand ((Square, State, Square), (Square, State, Square))

PreSMCommand ((Square, StateOmega), (Square, StateOmega))

data Square

This is a data type of square of the Turing machine TM.

Value is a square of string.

E is a square, which we use after change of Turing machine, for representing empties of a tape, and its argument marks the tape number.

RBS is a right bounding square.

LBS is a left bounding square.

ES is a empty square.

PCommand is a square of command with parentheses surroundings.

BCommand is a square of command with brackets surroundings.

Constructors

Value

Fields

val_name :: String

val_quote_cnt :: Int

E Int

RBS

LBS

ES

PCommand [TapeCommand]

BCommand [TapeCommand]

newtype InputAlphabet

Type of input alphabet of the Turing machine TM.

Constructors

InputAlphabet (Set Square)

newtype TapeAlphabet

Type of tape alphabet of the Turing machine TM.

Constructors

TapeAlphabet (Set Square)

newtype Commands

Type of commands of a Turing machine TM.

Constructors

Commands (Set [TapeCommand])

newtype TM

This type represents a Turing macine.

Constructors

TM (InputAlphabet, [TapeAlphabet], MultiTapeStates, Commands, StartStates, AccessStates)