Time

Description

Module Time include useful functions for working calculations which must be done after a certain period of time.

Documentation

data TimeAccuracy

Constructors

MS

S

M

H

class Monad m => MonadTimeout m

Minimal complete definition

timeout

timeout :: (MonadTimeout m, RealFrac i, NFData a) => i -> TimeAccuracy -> m a -> m a

timeout' :: RealFrac i => i -> TimeAccuracy -> IO a -> IO (Maybe a)