TmsParser

Description

This module provides types for parsing Tms Turing machines.

Documentation

parser :: Parser Tms

parseTms :: String -> String -> IO (Either String Tms)

Parse Tms from file.

pTms :: Parser Tms

Tms. Example:

name: TuringMachine_1 init: q1 accept: q2, q3

q1, _, x q2, a, y, -, >

q0, _, _ q3, a, _, <, - Note: lines should not end with ' ', also there could not be any newlines after Tms.