TmsType

Description

This module provides types for representing turing machines in a format suitable for the service: https://turingmachinesimulator.com/ Below this format will be called Tms.

Documentation

data TmsTapeSquare

Type of Tms tape square action.

Leave is leave any character unchanged.

'ChangeFromTo f t' is change it from f to t.

Constructors

Leave

ChangeFromTo Char Char

data TmsTapeHeadMovement

Type of Tms tape head movement

Constructors

MoveLeft

Stay

MoveRight

newtype TmsState

Type of Tms State.

Constructors

TmsState String

newtype TmsSingleTapeCommand

Type of Tms command for one tape. TmsSingleTapeCommand (action, movement).

Constructors

TmsSingleTapeCommand (TmsTapeSquare, TmsTapeHeadMovement)

type OneTapeTMCommand = (TmsState, TmsSingleTapeCommand, TmsState)

Type of Tms single tape command

toTmsCommand :: OneTapeTMCommand -> TmsCommand

newtype TmsCommand

Type of Tms command for entire Turing machine.

Constructors

TmsCommand (TmsState, [TmsSingleTapeCommand], TmsState)

newtype Tms

Type of Tms format. Tms (name, init accept commands, tapeAlphabets).

Constructors

Tms (String, TmsState, [TmsState], [TmsCommand], [String])

filterStateName :: String -> String

Process string so that it does not contain illegal characters.

fst3 :: (a, b, c) -> a

snd3 :: (a, b, c) -> b

thd3 :: (a, b, c) -> c