TuringMachine

Description

Module TuringMachine include type of Turing machine and useful objects for working with it. This module also export other useful modules.

Documentation

type LabeledStates = PrismMap String State

type Alphabet = IsoMap ShowedSymbol Symbol

data TuringMachine

turingMachine :: Quadruples -> LabeledStates -> Alphabet -> TuringMachine

quadruples :: Lens' TuringMachine Quadruples

labeledStates :: Lens' TuringMachine LabeledStates

alphabet :: Lens' TuringMachine Alphabet

type TMGetter a = Getter TuringMachine a

type TMSetter a b = Setter TuringMachine TuringMachine a b

states :: TMSetter State (Maybe State)

symbols :: TMSetter Symbol (Maybe Symbol)

allStates :: TMGetter (Set State)

allSymbols :: TMGetter (Set Symbol)

strQuadruples :: TMGetter (Set StrQuadruple)

strLabeledStates :: TMGetter (Set String)

strSymbols :: TMGetter (Set String)

strNumStates :: TMGetter (Set String)

strNumSymbols :: TMGetter (Set String)

maxNumState :: TMGetter Int

maxNumSymbol :: TMGetter Int

module TuringMachine.Quadruple

data Map k a

A Map from keys k to values a.

The Semigroup operation for Map is union, which prefers values from the left operand. If m1 maps a key k to a value a1, and m2 maps the same key to a different value a2, then their union m1 <> m2 maps k to a1.

data Set a

A set of values a.

newtype Pair a

Constructors

Pair

Fields

unPair :: (a, a)

unQuadruple :: Quadruple a -> (a, a, a, a)

data IsoMap a b

data PrismMap k a

class Deletable c v where

Minimal complete definition

Nothing

Methods

delete :: MonadFail m => v -> c -> m c

default delete :: (Containable c v, MonadFail m) => v -> c -> m c

unsafeDelete :: v -> c -> c

(\>) :: v -> c -> c infixr 7

(<\) :: c -> v -> c infixl 6

(<\~) :: ASetter s t c c -> v -> s -> t infixr 4

(<\=) :: MonadState s m => ASetter s s c c -> v -> m () infix 4

class Insertable c v where

Minimal complete definition

Nothing

Methods

insert :: MonadFail m => v -> c -> m c

default insert :: (Containable c v, MonadFail m) => v -> c -> m c

unsafeInsert :: v -> c -> c

(+>) :: v -> c -> c infixr 7

(<+) :: c -> v -> c infixl 6

(<+~) :: ASetter s t c c -> v -> s -> t infixr 4

(<+=) :: MonadState s m => ASetter s s c c -> v -> m () infix 4

class Filterable c v where

Minimal complete definition

filterC

Methods

filterC :: (v -> Bool) -> c -> c

(<?>) :: (v -> Bool) -> c -> c infix 4

(<?>~) :: ASetter s t c c -> (v -> Bool) -> s -> t infixr 4

(<?>=) :: MonadState s m => ASetter s s c c -> (v -> Bool) -> m () infix 4

class Gunctor c1 c2 v1 v2 | c1 c2 -> v1 v2 where

Minimal complete definition

Nothing

Methods

gmap :: (v1 -> v2) -> c1 -> c2

(<@>) :: (v1 -> v2) -> c1 -> c2 infix 4

class Containable c v where

Minimal complete definition

Nothing

Methods

member :: v -> c -> Bool

notMember :: v -> c -> Bool

class FastUnsafeListable c v | c -> v where

Methods

fromAscList :: [v] -> c

fromDistinctAscList :: [v] -> c

fromDescList :: [v] -> c

fromDistinctDescList :: [v] -> c

class UnsafeListable c v | c -> v where

Methods

fromList :: [v] -> c

class Listable c v | c -> v where

Minimal complete definition

Nothing

Methods

toList :: c -> [v]

default toList :: Valuable c v => c -> [v]

fromList_ :: MonadFail m => [v] -> m c

default fromList_ :: (UnsafeListable c v, MonadFail m) => [v] -> m c

class Operable c where

Minimal complete definition

union, intersection, difference

Methods

union :: c -> c -> c

(\/) :: c -> c -> c infixl 5

(\/~) :: ASetter s t c c -> c -> s -> t infixr 4

(\/=) :: MonadState s m => ASetter s s c c -> c -> m () infix 4

intersection :: c -> c -> c

(/\) :: c -> c -> c infixl 5

(/\~) :: ASetter s t c c -> c -> s -> t infixr 4

(/\=) :: MonadState s m => ASetter s s c c -> c -> m () infix 4

difference :: c -> c -> c

(\\) :: c -> c -> c infixl 5

(\\~) :: ASetter s t c c -> c -> s -> t infixr 4

(\\=) :: MonadState s m => ASetter s s c c -> c -> m () infix 4

class Slicable c k where

Minimal complete definition

takeC, dropC

Methods

takeC :: k -> c -> c

dropC :: k -> c -> c

splitAtC :: k -> c -> (c, c)

class Indexable c k v | c k -> v, c v -> k where

Minimal complete definition

Nothing

Methods

(!?) :: MonadFail m => c -> k -> m v infixl 9

(!<>) :: c -> k -> Set v infixl 9

(.@) :: MonadFail m => Getting x y c -> k -> Getting x y (m v) infixr 9

(.<@>) :: Getting x y c -> k -> Getting x y (Set v) infixr 9

class Keyable c k | c -> k where

Minimal complete definition

Nothing

Methods

keys :: c -> [k]

default keys :: Ord k => c -> [k]

keysSet :: Ord k => c -> Set k

class Valuable c v where

Minimal complete definition

Nothing

Methods

values :: c -> [v]

default values :: Ord v => c -> [v]

valuesSet :: Ord v => c -> Set v

class Singletonable c v | c -> v where

Methods

singleton :: v -> c

class Nullable c where

Minimal complete definition

emptyC

Methods

nullC :: c -> Bool

default nullC :: Sizable c => c -> Bool

emptyC :: c

class Sizable c where

Methods

size :: c -> Int

data Index

index :: Int -> Index

module Lens